Stage 6 - Functional
In this version, the code is fully functional. The user is able to bounce the ball around with the paddle and the ball responds correctly. In the next update I will be working on allowing the user to replay the game and also give stats for how long they have lasted. I will also be looking at allowing the ball to gain speed over time.